Keep track of visited vertices
When computing which vertices is reachable from a given vertex, I need
to keep track of what vertices has already been visited as I will
otherwise keep looking at the same vertices if there are cycles. I
therefore create a new property on a vertex that is simply an integer
indicating if it has been visited or not. Whenever this property is
used, it is assumed that it is reset afterwards to not visited.