Make vertex ID's unique across graphs
To allow for using multiple graphs and still distinguish between their
vertices, I make the vertex ID global. Furthermore the ID starts at one
which leaves the ID 0 to signal that some vertex is non existing or that
some error happened.

The change mean that I can no longer just index with the vertex ID into
the graph to get the correct vertex. I now have to loop through the
graph looking for the correct vertex.