Moved vertex responsibility
Before each graph had the responsibility of maintaining its own list of
vertices. As graphs were very liberal with its vertex pointers, this
lead to dangling pointers when the vertices were moved around in the
graph. I have made a centralised vertex store that can be asked for
pointers to a new vertex. Currently, there is no way to free a vertex
that has been claimed. I will see if this will become needed and
implement it if it is.

The change has given the graph the liberty to move around the vertices
locally in the graph as needed as the actual vertices are never moved in
the array of vertices.