Graph hold label data
Before each vertex had the responsibility for its label data. This had
some complications and involved each vertex holding a function pointer
that freed the resources used by its label. Now the graph is the holder
of the actual label data and each vertex contains a pointer to the
graphs array of labels.

I have disallowed a custom free function, so a label can no longer
contain anything that has been allocated on the heap.