Keep ID's of the old graph when copying
A user is probably suspecting that a copy of a graph that has a path
from vertex 1 to vertex 10 should also have a path from vertex 1 to
vertex 10. Before a copy had none of the vertices from the old graph, it
now still has new vertices, but the vertices retain their unique ID's
such that the same path can be found in both a graph and its copy.